requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-cosmos==3.0.2
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
django>=2.1.5
grpcio==1.14.2
grpcio-tools==1.14.2
gunicorn==19.6.0
idna==2.7
numpy==1.15.4
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1

